Add Fediverse handle functions and update changelog
